+++ src/routes/prompt/credential-offer/+page.svelte
@@ -12,8 +12,8 @@
   import Checkbox from '$src/lib/components/atoms/Checkbox.svelte';
   import Image from '$src/lib/components/atoms/Image.svelte';
   import PaddedIcon from '$src/lib/components/atoms/PaddedIcon.svelte';
+  import ListItemCard from '$src/lib/components/molecules/ListItemCard.svelte';
   import TopNavBar from '$src/lib/components/molecules/navigation/TopNavBar.svelte';
-  import ListItemCard from '$src/lib/credentials/ListItemCard.svelte';
   import { state } from '$src/stores';
 
   import Check from '~icons/ph/check-bold';
